fixed bgtz bug

+++ src/sim.c
@@ -292,7 +292,7 @@ void execute()
         // branch on greater than zero
         // if($s > 0) -> pc += i << 2
         case OP_BGTZ:
-            if(CURRENT_STATE.REGS[dcd_rs] > 0)
+            if((int32_t)CURRENT_STATE.REGS[dcd_rs] > 0)
                 NEXT_STATE.PC = CURRENT_STATE.PC + (dcd_se_imm << 2);
             break;
         // load upper immediate
